Deduplicate updates using a symbol
This addresses an issue: if you start by adding a simple atom and update
pair, everything works well in your component, but if you want to share
it with another component then it's impossible to tell which update is
authoritative. You can't let two updates change the same state on the
same signal.

Now updates are associated with an ID, allowing the store to recognize
two update handlers that represent the same operation and deduplicate
them.