Change sources type to RO arrays
This resolves an issue where sources were being inferred as polymorphic
lists instead of tuples. If you had two sources, one of a string and the
second a number, it would infer each payload as `string | number`. Now
you can assign a `[strAtom, nrAtom] as const` and it will infer the
correct types.

I'm not thrilled with that requirement. I might switch to objects if
this causes more trouble.